refactor: 👷 create new folder /props
