+++ components/layout/page-navigation/DefaultNoLinkNavigation.tsx
@@ -4,7 +4,7 @@ import IconArrow from '@/components/icons/IconArrow'
 
 import { NavigationDirectionEnum } from '@/lib/utils/typeDefinitions/enums'
 
-import { DefaultNoLinkNavigationProps } from '@/lib/utils/typeDefinitions/props'
+import { DefaultNoLinkNavigationProps } from '@/lib/utils/typeDefinitions/props/props'
 
 const DefaultNoLinkNavigation: FC<DefaultNoLinkNavigationProps> = ({ arrowDirection }) => {
   return (
