+++ components/layout/projects/project-page/CustomersCount.tsx
@@ -3,7 +3,7 @@ import { FC } from 'react'
 import List from '@/components/shared/List'
 import ListItem from '@/components/shared/ListItem'
 
-import { CustomersCountProps } from '@/lib/utils/typeDefinitions/props'
+import { CustomersCountProps } from '@/lib/utils/typeDefinitions/props/props'
 
 const CustomersCount: FC<CustomersCountProps> = ({ count }): JSX.Element => {
   return (
