+++ components/layout/projects/project-page/ProjectResponsibility.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import PageSection from '@/components/layout/projects/project-page/PageSection'
 
-import { ProjectResponsibilityProps } from '@/lib/utils/typeDefinitions/props'
+import { ProjectResponsibilityProps } from '@/lib/utils/typeDefinitions/props/props'
 
 const ProjectResponsibility: FC<ProjectResponsibilityProps> = ({ sections }): JSX.Element => {
   return (
