+++ components/pages/resume/ResumePlainTextPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import { ResumePlainTextPageProps } from '@/lib/utils/typeDefinitions/props'
+import { ResumePlainTextPageProps } from '@/lib/utils/typeDefinitions/props/props'
 
 const ResumePlainTextPage: FC<ResumePlainTextPageProps> = ({ content }): JSX.Element => {
   return (
