+++ components/shared/BreadcrumbsItem.tsx
@@ -3,7 +3,7 @@ import { FC } from 'react'
 
 import IconCaretRight from '@/components/icons/IconCaretRight'
 
-import { BreadcrumbsItemProps } from '@/lib/utils/typeDefinitions/props'
+import { BreadcrumbsItemProps } from '@/lib/utils/typeDefinitions/props/props'
 
 const BreadcrumbsItem: FC<BreadcrumbsItemProps> = ({ url, text }): JSX.Element => (
   <li>
