+++ components/shared/Heading2.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import { getDataTestIdAttribute } from '@/lib/utils/helpers/getDataTestIdAttribute'
 
-import { Heading2Props } from '@/lib/utils/typeDefinitions/props'
+import { Heading2Props } from '@/lib/utils/typeDefinitions/props/props'
 
 const Heading2: FC<Heading2Props> = ({
   children,
