+++ components/shared/SkillCard.tsx
@@ -4,7 +4,7 @@ import { FC } from 'react'
 import { getSkillCardID } from '@/lib/utils/helpers/skills/getSkillCardID'
 import { getYearsText } from '@/lib/utils/helpers/years/getYearsText'
 
-import { SkillCardProps } from '@/lib/utils/typeDefinitions/props'
+import { SkillCardProps } from '@/lib/utils/typeDefinitions/props/props'
 
 const SkillCard: FC<SkillCardProps> = ({ id, imgSrc, imgAlt, title, years }): JSX.Element => {
   return (
